pyproject.toml
[project]
name = "instabase-app-cicd-toolkit"
version = "0.0.2"
requires-python = ">=3.10"
authors = [
    { name = "Hannah Roiter", email = "hannah.roiter@instabase.com" },
    { name = "Ganan Prabaharan", email = "ganan.prabaharan@instabase.com" },
    { name = "Ben Hope", email = "ben.hope@instabase.com" },
]
description = "Tools for building CI/CD workflows Instabase Solutions and Apps."
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]
dependencies = [
  "requests",
  "python-dotenv"
]

[project.optional-dependencies]
test = [
  "black",
  "pytest",
  "pytest-cov",
  "pytest-sugar"
]

[project.scripts]
ib-cicd = "ib_cicd.promote_solution:main"
ib-cicd-solution-builder = "ib_cicd.promote_sb_solution:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
include = [
  "ib_cicd/",
]
